+++ src/utils/index.ts
@@ -2,5 +2,6 @@ export * from './aggregate';
 export * from './cache';
 export * from './dependency';
 export * from './fetch';
+export * from './isfunc';
 export * from './load';
 export * from './setup';